+++ curated_transformers/layers/__init__.py
@@ -15,6 +15,6 @@ from .embeddings import (
     RotaryEmbeddings,
     SinusoidalPositionalEmbedding,
 )
-from .encoder import EncoderLayer
 from .normalization import RMSNorm
 from .scalar_weight import ScalarWeight
+from .transformer import DecoderLayer, EncoderLayer, TransformerLayerNorms
