+++ curated_transformers/models/gpt_neox/__init__.py
@@ -1,4 +1,3 @@
 from .causal_lm import GPTNeoXCausalLM
 from .config import GPTNeoXAttentionConfig
 from .decoder import GPTNeoXDecoder
-from .layer import GPTNeoXDecoderLayer
