+++ curated_transformers/models/gpt_neox/_hf.py
@@ -76,8 +76,8 @@ def convert_hf_state_dict(cls, params: Mapping[str, Tensor]) -> Mapping[str, Ten
         name = re.sub(r"\.dense_4h_to_h", r".output", name)
 
         # Layer norms
-        name = re.sub(r"\.input_layernorm", r".input_layer_norm", name)
-        name = re.sub(r"\.post_attention_layernorm", r".ffn_layer_norm", name)
+        name = re.sub(r"\.input_layernorm", r".attn_input_layer_norm", name)
+        name = re.sub(r"\.post_attention_layernorm", r".ffn_input_layer_norm", name)
         name = re.sub(r"final_layer_norm\.", r"output_layer_norm.", name)
 
         # Embeddings
