+++ curated_transformers/models/gpt_neox/decoder.py
@@ -1,17 +1,23 @@
+from functools import partial
 from typing import Any, List, Mapping, Optional, Type, TypeVar
 
 import torch
 from torch import Tensor
 from torch.nn import Dropout, Embedding, LayerNorm, ModuleList
 
-from ...layers.attention import AttentionMask
+from ...layers.attention import (
+    AttentionMask,
+    QkvHeadSharing,
+    QkvMode,
+    RotaryEmbeddingConfig,
+)
 from ...layers.cache import KeyValueCache
+from ...layers.transformer import DecoderLayer, TransformerLayerNorms
 from ..hf_hub import FromHFHub
 from ..module import DecoderModule
 from ..output import ModelOutputWithCache
 from ._hf import convert_hf_config, convert_hf_state_dict
 from .config import GPTNeoXConfig
-from .layer import GPTNeoXDecoderLayer
 
 # Only provided as typing.Self in Python 3.11+.
 Self = TypeVar("Self", bound="GPTNeoXDecoder")
@@ -44,9 +50,36 @@ class GPTNeoXDecoder(DecoderModule, FromHFHub):
         )
         self.dropout = Dropout(p=config.embedding.dropout_prob)
 
+        layer_norm = partial(
+            LayerNorm,
+            config.layer.hidden_width,
+            config.layer.layer_norm_eps,
+            device=device,
+        )
         self.layers = ModuleList(
             [
-                GPTNeoXDecoderLayer(config.layer, config.attention, device=device)
+                DecoderLayer(
+                    attention_dropout=config.attention.dropout_prob,
+                    hidden_act=config.layer.hidden_act,
+                    hidden_dropout=config.layer.dropout_prob,
+                    hidden_width=config.layer.hidden_width,
+                    intermediate_width=config.layer.intermediate_width,
+                    layer_norms=TransformerLayerNorms(
+                        attn_input_layer_norm=layer_norm(),
+                        ffn_input_layer_norm=layer_norm(),
+                    ),
+                    num_attention_heads=config.attention.num_attention_heads,
+                    parallel_attention=True,
+                    qkv_head_sharing=QkvHeadSharing.NONE,
+                    qkv_mode=QkvMode.MERGED_SPLIT_BEFORE,
+                    rotary_embeds=RotaryEmbeddingConfig(
+                        fraction=config.attention.rotary_fraction,
+                        base=config.attention.rotary_base,
+                    ),
+                    use_bias=True,
+                    use_gate=False,
+                    device=device,
+                )
                 for _ in range(config.layer.num_hidden_layers)
             ]
         )
