+++ curated_transformers/models/llama/__init__.py
@@ -1,4 +1,3 @@
 from .causal_lm import LLaMACausalLM
 from .config import LLaMAConfig
 from .decoder import LLaMADecoder
-from .layer import LLaMADecoderLayer
