+++ curated_transformers/tests/generation/test_dolly_v2.py
@@ -25,8 +25,8 @@ def test_generate_deterministic(dolly_generator):
         "What is spaCy?",
     ]
     answers = [
-        "Rust is a multi-paradigm, high-level, general-purpose programming language. Rust is designed to have a small, stable, and fast implementation. Rust is also designed to be easy to learn. Rust is intended to be used for writing fast, reliable, and portable code.\n\n",
-        "SpaCy is an open-source natural language processing (NLP) library for Python. It is designed to be fast, scalable, and easy to use.\n\n",
+        "Rust is a multi-paradigm, high-level, general-purpose programming language. Rust is designed to have a small, consistent, and predictable language surface. Rust is also designed to be efficient, and to have a small memory footprint. Rust is designed to be safe, and to have a well-defined memory model. Rust is also designed to be concurrent, and to have a good support for concurrent programming. Rust is designed to be fast, and to have a good support for performance-critical code. Rust is also designed to be modular, and to have a good support for modular programming. Rust is designed to have a good support for internationalization and localization.\n\n",
+        "SpaCy is a natural language processing (NLP) library for Python that provides tokenization, part-of-speech (POS) tagging, named entity recognition (NER), and dependency parsing.\n\n",
     ]
     assert dolly_generator(prompts, config=GreedyGeneratorConfig()) == answers
 
@@ -44,10 +44,8 @@ def test_generate_max_generated_pieces(dolly_generator):
         "What is the Rust programming language?",
         "What is spaCy?",
     ]
-    answers = [
-        "Rust is a multi-paradigm,",
-        "SpaCy is an open-source natural language",
-    ]
+    answers = ["Rust is a multi-paradigm,", "SpaCy is a natural language processing (N"]
+
     assert (
         dolly_generator(prompts, config=GreedyGeneratorConfig(max_generated_pieces=10))
         == answers
@@ -73,14 +71,14 @@ def test_generate_sample(dolly_generator):
     # Fix the seed so that we are always randomly sampling in the same way.
     torch.manual_seed(0)
     assert dolly_generator(prompts, config=SampleGeneratorConfig(top_k=10)) == [
-        "SpaCy (short for Spanish-Portuguese-Chinese artificial intelligence language model) is an open-source language model (LLM) that uses a novel combination of deep learning, big-data, and semantic parsing to provide the best accuracy available with no human involvement. It is one of the most effective open-source language models.\n\n",
-        "SpaCy is an open-source package for Python that provides language model (LM) and automatic speech recognition (ASR) functionality.\n\n",
+        "SpaCy (short for spaCy Toolkit) is a Python library for Natural Language Processing (NLP) and machine translation based on the   spaCy research project. It has been open-source since June 2023 and has been used in production for NLP tasks like POS tag classification (part of OpenStax' TACL), question answering (QACQUEL), and semantic parsing (CoreNLP) as of October 2023.\n\n",
+        "SpaCy is an open-source package for Python that automates part of the process of building language models (LMs), or parsers for plain text, using a technique called graph-based learning. It supports English, French, German, Dutch, Italian, Spanish, Portuguese, Polish, and Dutch, all of which are official language groups of Europe.\n\n",
     ]
 
     torch.manual_seed(0)
     assert dolly_generator(
         prompts, config=SampleGeneratorConfig(top_k=5, temperature=2)
     ) == [
-        "SpaCy (short for spaCoN-based AI for Cyber-Security) is open source software that is used for automated language and named entity recognition (NER). SpaCy has been specifically developed for cybersecurity text processing and uses techniques that were developed for NLP over many years by the University of Pennsylvania's Named Entity Tagging (NetNesp) project, by Penn Tree Street Labs and other contributors, with a focus towards named- Entity Resolution (ner).\n\n",
-        "The spaCy library is an NLP package based on Stanford's SpanTweaked architecture for Stanford CoreNLP. SpaCy's main difference over Stanford CoreNLP is the usage of the PySpREnd language model, which was designed for Natural Language Processings, rather than just a simple parser, like the original SpanTweaked. This makes it possible to do much better NLP with the model than just the simple parsing of Stanford's ParserCore. However, spaCy is also a more general-purpose library, which also allows for more complex NLP tasks than just sentence parsing, such as part-of-speech taggers and dependency parsing, so it can be a good starting point for a NLTK-style library. In the field of machine learning and artificial intelligence, a spa file (spatial Pyramids file) describes a specific model trained on data, such as a part-of-speech tagged corpus, and it can be loaded and executed with spaCy's model or the Stanford CoreNLP's models (see Stanford CoreNLIps's API for loading a model from a spaCy spa file).\n\n",
+        "SpaCy (short for Spanish Language Model) is a natural language processor (NLP) for English based on the open source spaCy project. It supports sentence detection, part-of-speech tagging and coreference chains, among other things. spaCy was originally developed at the University of Cambridge's Machine Translation Lab. Since then, it has also developed a version optimized for Python that can be installed viapip.\n\n",
+        "The spaCy library is an NLP library based on Stanford's PoSpell architecture, that is designed to make NLP easy. spaCy supports all the standard NLTK pipeline stages and is able to outperform the NLTK on many Named Entity Tagging tasks, as well many others, both at test-level and at the system-wide average. It has a similar architecture to NLTK’s, but is designed from the ground up for the needs of a NLP research community rather than a production system: It has a smaller and less mature API, does away with its core tokenizer (which is notoriously hard to train), and is based on the Speller system from Stanford's CS-ADLDN program, which has been shown to significantly outperform the NLTK tokeniser in terms of both accuracy (93.7% on a test set of 20K tokens vs. NLTK's 76.3%) and efficiency.\n\n",
     ]
